Remove reference to the company's location
MSCD 2.68 recommends against including the address of a party that is a registered legal entity, because the combination of the full official name + the place in which the organization is registered should be sufficient to uniquely identify the organization, and so the additional information is clutter. See also MSCD 1.41-1.58, "Contract Language Should Omit Redundancy".