Add a date
MSCD discusses dates in some depth in 2.21-2.45 and 5.3-5.7. A variety of options are discussed, such as what one might want to do in the situation where an electronic signing service is used. This commit gives the agreement a date (and related signature block) in accordance with MSCD's recommended defaults, but users of the agreement might want to consult MSCD and amend the agreement appropriately to their situation.

Note that this removes "Agreed to and accepted." I find this simple phrase much easier on the eyes than MSCD's recommended, wordier phrase, which is unusual - MSCD's recommendations usually, to my eye, improve readability. However, (1) I'm seeking in this experiment to follow MSCD's recommendations, not necessarily my own instincts; and (2) MSCD's recommendations in this area seem well justified - you can find them at 5.8-5.22.

[Note that, in the original intended use case of this document, the agreement may be dated in several other ways - for example, by submission to the patent office. However, as it is usually good form to date stand-alone agreements, I've included these terms here, whether or not strictly necessary.]