remove a "hereof"
MSCD hits on hereof in three places - 4.92, 13.260, and 15.12; the relevant one here is 13.260, which simply recommends avoiding all "here-" and "there-" words, correctly calling them "one of the hallmarks of legalese." (15.12 admits that there may be situations where here-/there- are useful for concision, but tellingly, no examples are offered.)

Relatedly, MSCD 2.28 recommends using "the date of this agreement" rather than the traditional "Effective Date" for simplicity and readability. So I have done that here.