Fix the "lead-in"
This is probably the first change that will make lawyers nervous, specifically because of the removal of the recital of consideration. 

In a passage that is typical of the book, 2.149-2.164 of MSCD goes into some detail about why the "recital of consideration" is unnecessary and should be replaced by more specific statements about what the consideration actually is. The discussion is not merely his opinion, but cites the Restatements and some caselaw; while also lecturing the courts on what they should do (a behavior that I find tiresome but which is of course common among treatise-writers).