do hereby -> hereby
MSCD 3.23: "do hereby" is an archaism, so kill it. Again, an example where MSCD is incredibly detail oriented. Note, though, that "hereby" is kept, despite being a previously attacked "here-" word. That's because hereby is (to Ken) the best way to express "language of performance", which this phrase is. 

Again, this is Ken's knack for, and love of, categorization shining through. Chapter 3 of the book is about "categories of contract language", where different categories (like language of agreement, of obligation, and of performance) are identified, and rules (and examples) set out for each one. In this case, no less than 16 different examples of language of performance are identified, and they are graded from "avoid" to "not recommended" to "could be improved on" to "recommended." "Hereby", in this context, comes out on top, and once you've read his analysis, you'll probably agree.

[Again, trying to stick to edits MSCD specifically recommends. That said, here, the mixed-voice (mixed-tense?) 'sell'/'have sold' language is very unclear and unusual, and could probably stand to be rethought or removed.]