Correctly capitalize "inventor".
This fixes the capitalization of "inventor" in Sec. 4. 

This sentence uses both "Inventor" and "Inventors", though only "Inventors" is defined. I have not changed that, because MSCD, in the chapter on Defined Terms, recommends against the common practice of redundantly defining singular and plural terms (like "'Inventor' or 'Inventors' means...") on the sensible ground that the less redundant definition "wouldn't confuse a reasonable reader."