Make section tabulation consistent
There are places in MSCD where the rules are less clear than others; the section on enumerated clauses (MSCD 4.28-4.37) is among them. In this case I have added a missing colon, and otherwise simply made the formatting consistent with the next section. 

MSCD also recommends using (1), (2)... instead of (a), (b)... here, but I find the rationale particularly confusing and uncompelling, so (unlike most of the rest of this review) I have ignored that recommendation.