Switch from represents to states
One of the categories of contract language that MSCD goes into in some depth is "language of declaration" - i.e., language used by the parties to declare facts, a category further split into statements - used when one party wants to assert that a statement is accurate - and acknowledgements - used when one party wants the other to admit knowledge and accuracy of a statement.

Lawyers traditionally use "represents and warrants" to handle statements, but MSCD goes on at great length (citing caselaw, the UCC, and a few dictionaries) as to why this is a bad idea, and why "states" should be used instead. Importantly, he emphasizes that relying on "represents" to trigger a specific sort of remedy is unreliable, and that instead contract drafters should be explicit about what remedies they expect and how they will be enforced. Changing it here is a reminder that the book will make some lawyers uncomfortable (by challenging ritual word use) and can help drafting (by encouraging drafters to be precise and explicit about things like remedies).